feat(#1457): [Architecture] Split oversized cmd/cheasee-pi files, break auth/config/settings import cycle
